Cutoff SDL event loop
